requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
